backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
docker==7.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
slowapi==0.1.9
anthropic==0.39.0

# Testing
# Note: httpx is also used for GitHub API calls
pytest==8.0.0
pytest-asyncio==0.23.0
httpx==0.26.0
pytest-cov==4.1.0
